+++ src/types.ts
@@ -42,8 +42,9 @@ export interface BindingOptions<T> {
  */
 export interface ListOptions<T> {
   key: keyof T | ((item: T, index: number) => string | number);
-  render: (item: T, index: number) => string;
+  render: (item: T, index: number) => string | Element;
   bind?: ($el: JQuery, item: T, index: number) => void;
+  update?: ($el: JQuery, item: T, index: number) => void;
   onAdd?: ($el: JQuery) => void;
   onRemove?: ($el: JQuery) => Promise<void> | void;
   empty?: string;